+++ packages/core/src/action/index.ts
@@ -1,8 +1,26 @@
 import { z } from 'zod';
 
-import { Mastra } from '../mastra';
+import { Agent } from '../agent';
+import { MastraEngine } from '../engine';
+import { LLM } from '../llm';
+import { ModelConfig } from '../llm/types';
+import { Logger } from '../logger';
+import { MastraMemory } from '../memory';
+import { SyncAction } from '../sync';
+import { Telemetry } from '../telemetry';
+import { MastraVector } from '../vector';
 import { WorkflowContext } from '../workflows';
 
+export type MastraPrimitives = {
+  logger?: Logger;
+  telemetry?: Telemetry;
+  engine?: MastraEngine;
+  agents?: Record<string, Agent>;
+  vectors?: Record<string, MastraVector>;
+  memory?: MastraMemory;
+  syncs?: Record<string, SyncAction<any, any, any, any>>;
+  llm?: (model: ModelConfig) => LLM;
+};
 export interface IExecutionContext<
   TSchemaIn extends z.ZodSchema | undefined = undefined,
   TContext extends WorkflowContext = WorkflowContext,
@@ -11,9 +29,8 @@ export interface IExecutionContext<
     ? z.infer<TSchemaIn> & { machineContext?: TContext }
     : { machineContext?: TContext };
   runId?: string;
-  mastra?: Mastra;
+  mastra?: MastraPrimitives;
 }
-
 export interface IAction<
   TId extends string,
   TSchemaIn extends z.ZodSchema | undefined,
@@ -24,7 +41,7 @@ export interface IAction<
   description?: string;
   inputSchema?: TSchemaIn;
   outputSchema?: TSchemaOut;
-  mastra?: Mastra;
+  mastra?: MastraPrimitives;
   payload?: TSchemaIn extends z.ZodSchema ? Partial<z.infer<TSchemaIn>> : unknown;
   execute: (context: TContext) => Promise<TSchemaOut extends z.ZodSchema ? z.infer<TSchemaOut> : unknown>;
   [key: string]: any;