+++ examples/vnext-playground/CHANGELOG.md
@@ -1,5 +1,18 @@
 # examples-vnext-playground
 
+## 1.0.1-alpha.12
+
+### Patch Changes
+
+- Updated dependencies [f537e33]
+- Updated dependencies [bc40916]
+- Updated dependencies [f7d1131]
+- Updated dependencies [75bf3f0]
+- Updated dependencies [3c4488b]
+- Updated dependencies [d38f7a6]
+  - @mastra/core@0.1.27-alpha.36
+  - @mastra/engine@0.0.5-alpha.30
+
 ## 1.0.1-alpha.11
 
 ### Patch Changes